pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-cohere>=0.5.1",
    "langchain-groq>=1.1.2",
    "langchain-pinecone>=0.2.13",
    "pinecone-client>=6.0.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.44.0",
]
